obj files can be loaded into vertex and index buffers

+++ engine/src/renderer/vulkan/vulkan_defines.h
@@ -74,7 +74,7 @@ typedef struct VulkanContext {
   VkSemaphore* render_complete_semaphores;
   VkFence draw_fences[MAX_FRAMES_IN_FLIGHT];
 
-  u64 vertex_count;
+  u32 vertex_count;
   Vertex* vertices;
   VkVertexInputBindingDescription vertex_binding_description;
   VkBuffer staging_vertex_buffer;
@@ -82,7 +82,7 @@ typedef struct VulkanContext {
   VkBuffer vertex_buffer;
   VkDeviceMemory vertex_buffer_memory;
 
-  u64 index_count;
+  u32 index_count;
   u32* indices;
   VkBuffer staging_index_buffer;
   VkDeviceMemory staging_index_buffer_mem;
